Store the offset of a rule/character/CharDotsMapping in the struct
Getting the TranslationTableOffset of a TranslationTableRule,
TranslationTableCharacter or CharDotsMapping is similar to getting the
memory address, but unlike the address the offset is a reliable way to
get the data because the location of the table in memory may change
several times during compilation. Offsets don't change.

Making the offset a field of the object itself simplifies the code
because it reduces redundancy in the arguments being passed around.

Also discovered some potential causes of segmentation faults and fixed
them.